Make the directories metadata generation script 70% faster.

### Comparative table

| Folder | #&nbsp;subfolders | Before&nbsp;<small>(s)</small> | <small>per&nbsp;subdir&nbsp;(ms)</small> | After&nbsp;<small>(s)</small> | <small>per&nbsp;subdir&nbsp;(ms)</small> | Δ |
|:-|:-:|-:|-:|-:|-:|-:|
| `/assets/folder-a/asset-a` | 10 | `2.199` | `220` | `0.661` | `66` | <mark>-70%</mark> |
| `/assets/folder-a` | 286 | `76.377` | `267` | `20.813` | `73` | <mark>-73%</mark> |
| `/assets` | 1 435 | `637.707` | `444` | `175.573` | `122` | <mark>-72%</mark> |